single processing of a file is completed
implemented functions in find_bookless_reviews.php to read a file and
then process each line, removing the trailing comma and whitespace
characters --> json decoding it and then checking the URL to see if it
has any data associated with it.

In addition, the script writes all URLs with missing output to a file.