Changes!
Things have been moved, and minimized from find_bookless_reviews.
Namely, functions are now streamlined and centralized in
json_functions.php

The new file that we're working with is get_review_information() which
is a file that attempts to get the author/title/etc. from all the urls
stored in all the files in the Data directory